requirements.txt
requests==2.32.3
pymysql==1.1.1
jsonpath==0.82.2
pytest==8.3.2
openpyxl==3.1.5
allure-pytest==2.13.5
jinja2==3.1.4